Bug fixes for adding a sandbox.. Should not be able to add Sandbox if already exists due to overwrite.
